Use shared secret and receipt field (#24)
* Adopt getReceiptFieldValue for apple in_app
  If field is not found in object root, check in_app for value
* Use latest_receipt_info when available

latest receipt info contains up-to-date info on a receipt's history. Especially
regarding monthly accounts that auto-renew. We should provide the transaction id
and product id from this field when available. By default we use the receipt
in_app as a starting point.

* Reference latestReceiptInfo for apple receipts

Added a brief example that includes a monthly auto-renewable subscription as
well as a reference to when the in_app and latestReceiptInfo may be available.

* Sort latest receipt info by transaction id

transaction id is iterable and thus is a better field to sort by versus date