[Apple] Check for cancellation_date_ms first (#50)
[Apple] Check for cancellation_date_ms first

If cancellation date is available it means that Apple Customer Support refunded subscription payment.

From Apple:
To check whether a purchase has been canceled by Apple Customer Support, look for the
Cancellation Date field in the receipt. If the field contains a date, regardless of the
subscription’s expiration date, the purchase has been canceled. With respect to providing
content or service, treat a canceled transaction the same as if no purchase had ever been made.